Bump lib9c 1.7.2

+++ MarketService/Models/SkillModel.cs
@@ -13,7 +13,7 @@ public class SkillModel: ISkillSchema, ISkillModel
     public SkillCategory SkillCategory { get; set; }
     public int HitCount { get; set; }
     public int Cooldown { get; set; }
-    public int Power { get; set; }
+    public long Power { get; set; }
     public int StatPowerRatio { get; set; }
     public int Chance { get; set; }
     public StatType ReferencedStatType { get; set; }
